Simplify and use only one result page
